Add russh
Summary:
Adds crate `russh`, an SSH implementation in pure Rust.

This will be used in CoreSSH, as the ultimate goal of `fleet_exec`/`ssh_infra` is to reduce the proliferation of disparate SSH client usage - that way we can provide a level of predictability and support to ssh@meta - hence the limited visibility seen in `fixups.toml`.

Also note that the `disallowed_crates.bzl` has changed to now allow this. I am from the team it mentions to contact, and this is expected.

Reviewed By: edward-shen

Differential Revision: D62660825

fbshipit-source-id: d5c9a841adaa5d00531354c0f7cf217c738c4975