Label-based function hover for methods
Summary: The diff D62496510 produces nice types when hovering on top-level function calls and calls to static methods, but not instance methods. That's because the hover logic picks up the type on the `Id` bit of the `Obj_get` node; so we also need to transform this type in the same way that the type of the outer `func` field of the `Call` node has been transformed.

Reviewed By: madgen

Differential Revision: D62711238

fbshipit-source-id: ce946e527b4449a25c1a89963362e9c799889c5e