Don't check parameter type coercion
Summary:
We were checking that parameter type hints don't clash with any other binding's type in the environment.

This is not needed since rebinding a parameter is already a parse error and parameters would shadow any existing binding, so there's no need to check subtyping.

This diff eliminates the check and hence we can now typecheck the following program without a nonsensical expected string but got int error.
```
(string $_, int $_) ==> {}
```

Reviewed By: andrewjkennedy

Differential Revision: D62653615

fbshipit-source-id: caaee4213c647b18e36a9cb1edbb897232272100