Relax a DCHECK in the cursor serialization implementation to account for a streaming bug
Summary:
Due to a bug somewhere with streaming, the packet sent by the server ends up with an extra 0x00 byte at the end when received by the client, causing the DCHECK to fail.

Relaxing the DCHECK for now to unblock TAO/Sql over thrift usage of cursors. This relaxation of the check was discussed and agreed on offline, while we separately investigate and fix the extra byte bug.

Reviewed By: robertroeser

Differential Revision: D62780896

fbshipit-source-id: 456c9ca6e54933422bae4b2382e40314aba4597b