Add flag to disable format-on-save via `textDocument/willSaveWaitUntil`
Summary:
## Context

Around 8-10% Hack diffs have format issues on CI. For an integrated ecosystem like WWW it's too much. I believe there is a misbehaving local tooling involved which make it happen. As we're rolling out Linttool LSP for other languages we see improvements across the board. https://fb.workplace.com/groups/lintteam/permalink/1444481879599554/

So I'd like to conduct an experiment if swapping to Linttool LSP would prove a stat significant decrease in diffs with format issues. The rough plan is to have a GK that would

1) disable `textDocument/willSaveWaitUntil` for Hack LSP based on a given GK. To ensure that it doesn't interfere with linttool
2) on the linttool lsp side we'll use `hackfmt` binary to run formatter twice (as we know it's not idempotent)
3) We'll hookup `textDocument/willSaveWaitUntil` powered by Linttool LSP for users that pass a GK
4) Run experiment and see if it helps

Note: Current Hack LSP is also hooked up in Bento. We not gonna change this but Bento doesn't affect CI in any way so it can continue using Hack LSP as general LSP for now

Reviewed By: mheiber

Differential Revision: D62939417

fbshipit-source-id: 87831566a8adb8fc734afee5533fb214ecca9d31