adding test for modifying env within post-fork callback
Summary:
Will use this technique in the next diff in order to create a subprocess that contains an environment variable whose value is that subprocess's own pid.

It relies crucially on the environment strings remaining where they were when they were passed into the Subprocess c'tor (and not being moved or copied), so this test makes sure that that won't change on us.

Reviewed By: dmpolukhin

Differential Revision: D62691897

fbshipit-source-id: ced30dfb2daa30067958a7feea819d8e49b35d4b