Fix idempotence bug
Summary:
Hack formatter is supposed to be idempotent, i.e., there is no difference between invoking it once vs many times.

There was a bug because we preserve line breaks for collection literals even if we can in principle squash it in a single line.

This works as intended except for the empty literal case where we do (and want to) squash it onto one line, but the code explicitly looks for a newline character at the end of the left brace.

So on the first run we squash it into one line and the absence of the newline character leads to causes it to format further on the second run.

Reviewed By: mheiber

Differential Revision: D62971173

fbshipit-source-id: e469329ad1a2d7c68cb9fe12d5c6500f0036095a