Open tuples: type representation
Summary:
This is the first diff in a stack that implements tuples with *open tuples*. By analogy with function parameters, open tuples have *required* components, *optional* components, and a *variadic* component.

The type `(int,optional string, optional bool)` is a tuple of exactly 1, 2, or 3 elements. The first has type `int`, the second (if it exists) has type `string`, the third (if it exists) has type `bool`.

The type `(int,mixed...)` is a tuple of 1 or more elements, the first has type `int`, remaining elements have type `mixed`. (We might support the syntax `(int,...)` but for now let's be cautious and require an explicit type, as for variadic functions.)

The type `(optional int,string...)` is a tuple of 0 or more elements, the first (if it exists) has type `int`, the remaining elements (if they exist) have type `string`.

Optional and variadic elements are subject to the same restrictions as with optional and variadic parameters in functions:
* Optional components may not precede required components.
* A tuple may have only a single variadic component, and it must appear last in the tuple.

Internally, a closed tuple is represented by an open tuple whose variadic type is `nothing`.

This initial diff just introduces the internal representation of tuple types. There is no support for parsing, but some of the type-checker implementation (such as pretty printing) is included. Where functionality has to be filled in, there are TODO comments with task numbers.

Reviewed By: madgen

Differential Revision: D55472213

fbshipit-source-id: 6b40fd2d02014fb86564a40d1cb69ba13636cc83