Convert `--no_resource_pile_or_cc` flag to `--work_stealing_executor_only` in StressTestServer
Summary:
The `--no_resource_pile_or_cc` flag was created to show that we could create a resource pool with only an executor. More specifically, it is for testing with the work stealing executor that is now included in StressTestServer. Converting the flag to `--work_stealing_executor_only` makes the intent more clear. It does the same thing as the `--work_stealing` flag, but does not create a request pile or concurrency controller for the resource pool.

Additional cleanups:
- Add text to `--work_stealing` flag hint to clarify that it implies `--resource_pools`.
- Consolidate logic around resource pool creation.

Reviewed By: robertroeser

Differential Revision: D62513553

fbshipit-source-id: 353adfab1d05d0542b05aabdc061865e082eb71b