Use environment to store subtype relationships
Summary:
This diff introduces a subtyping map in the environment to keep track of subtypes.

This diff is a no-op, but we'll use it later to keep track of subtyping relationships such as those relating to enums (between `arraykey` and the enum for example). Currently, this is not possible and we use the type itself for classes to store the class hierarchy.

Reviewed By: mheiber

Differential Revision: D63235031

fbshipit-source-id: 0aca05bfcd089d6cde6cbb10e052360e1d772eff