Do not dereference structs on read/decode
Summary:
The goal of this diff is to improve the build speed of the generated code by simplifying Golang escape analysis.

Post about the issue: https://fb.workplace.com/groups/codegophers/permalink/26214126538209280/

### Long story:
I think I made a bad design decision in the implementation of the generator by using **values** rather than **pointers** when dealing with *structs* in the generator.

As a result there is a lot of dereferencing and address-taking in the generated code. I have a suspicion this is tripping up escape analysis - because things go from value to pointer and back between function calls - for no good reason. I imagine escape analysis becomes much more complex as a result and takes a very long time.

Fortunately - the bad decision is relatively easy to revert and fix. Pointers are now used by default when dealing with structs. Values (via dereference `*`) are only used in a couple of special places that require values.

Reviewed By: awalterschulze

Differential Revision: D57308440

fbshipit-source-id: ea0e610a8487feb226624fe517402bbb33eab567