Remove dead code in lambda typing (2)
Summary:
When checking lambdas with explicit type hints on lambdas, we localize the type hint twice: once when constructing the "function type" from the expected type and parameter info, and then a second time when calling `closure_bind_param`. This second localization is unnecessary as we already pass in the localized parameter type from the constructed function type.

Added a new test to make sure that we're plumbing the correct parameter type through to `closure_bind_param`, where expected type and explicit parameter type are in a subtype relationship.

Reviewed By: mheiber

Differential Revision: D63260727

fbshipit-source-id: 38591ecd45f2fd215d1ca7a90197d45905c01994