Move logic to assign blocks to assign blocks to code areas when hot/cold splitting is disabled
Summary: Move the logic to put all the blocks in the main code area when hot/cold splitting is disabled, to the same point where hot/cold splitting was performed. This was being done a bit too late and, in the rare case that the unit entry block is initially put in the cold area, an assert would fail before the entry was moved to main.

Reviewed By: mdko

Differential Revision: D63355475

fbshipit-source-id: f30300dcb75d503d69b9476ea8e8d6ba30f55f50