Fix handling for classname<T> enum bases
Summary:
Enums with base `classname<T>` are currently handled incorrectly. When resolving type-constraints in sandbox mode they are treated as `arraykey` while hhbbc correctly treats them as the `Classname` type.

Their handling for reified generics is also broken but because it was consistent between sandbox and repo mode I have left it alone to avoid introducing a breaking change here. It should be fixed as a follow-up.

Reviewed By: ricklavoie

Differential Revision: D63399497

fbshipit-source-id: 0ca77faf41640d8b71b876e895f5212c048ce54e