Remove ExtendedHint flag from TypeConstraint
Summary: We had been applying the `ExtendedHint` flag inconsistently when resolving types both in hhbbc and in the runtime. It originally existed for PHP compatibility and basically just implies that the Soft or Nullable flags are set now so it appears to be entirely redundant. This diff simply removes it.

Reviewed By: ricklavoie

Differential Revision: D63399502

fbshipit-source-id: b39d731da9fe1fbee30613aba3d1435e5f13e1e1