Fix type structure resolution for typevars field
Summary:
The `typevars` field is removed when we substitute in actual types during the resolution of a type structure. Where we were stripping it had been handled inconsistently between hhbbc and the runtime type structure helpers.

Interestingly I found that the way we handle `typename<T>` in hackc causes unbound type variables. We treat `typename<T>` as `typename` when emitting type structures in hackc effectively stripping any generic parameters. This may merit some followup but for now this diff handles this case as well by dropping the extra `typevars` field.

Reviewed By: ricklavoie

Differential Revision: D63399500

fbshipit-source-id: 330a0aeb891063f178f86aace6afc37e024bf844