Fix typename resolution for members of unions
Summary: This resolution was handled differently between hhbbc and the runtime. The only practical difference here seems to be whether or not the `typename` field is propagated in members of a union. We don't currently have a way to introspect this so it isn't a behavior altering change but for the purposes of ensuring identical resolution behavior it's worth fixing.

Reviewed By: ricklavoie

Differential Revision: D63399499

fbshipit-source-id: 869142df4ef5f20554e17d9990e327fa6032457c