Reinstate `if_unsat` combinator use for denotable unions
Summary:
When we have a denotable union in subtype position the behavior of extended reasons differs from the legacy error reporting.

In the legacy error, we use the `if_unsat` combinator to report the error associated with the union rather than at its elements. In the extended reasons error, we will generate errors for each of the elements.

Since we'd like to enable extended reasons by default and the current behavior is well established and well motivated, this diff reinstates the use with extended reasons.

Reviewed By: andrewjkennedy

Differential Revision: D63389803

fbshipit-source-id: e699210c281b6025a752d3038e509d91658b9b7d