Type splat: HackC changes
Summary:
Variadic parameters are implemented in HackC and HHVM by interpreting the parameter in the body of the function as a `varray`. The variadic calling convention tells HHVM to bundle up the actual arguments as an array.

This is exactly what we want for type-splat parameters too, as tuples are represented as arrays in HHVM. So it's really easy to provide support for type-splats in HHVM: simply set the variadic calling convention.

Differential Revision: D63387073

fbshipit-source-id: 3a59eee7c9e502c9c951641b4cacc354d2872303