Avoid logging tons of warnings, and simplify error logging
Summary:
Currently, the logs have tons of warnings. That's because we log errors before filtering out the warnings. In this diff, I move the error logging to near where we stream errors.

Previously, we also had an advanced way of capping the number of errors logged, which obviously didn't work given the tons of warnings I've seen being logged. I'm simplifying the capping and making it clear in the logs that things are capped.

Reviewed By: madgen

Differential Revision: D63392603

fbshipit-source-id: e8c8ca7e35c0df8c779a0d5217faf5b64c0efa63