Change Implementaion of SetRelation
Summary:
Currently a `SetRelation.t` is represented as:
```
type t =
  | Equal
  | Subset
  | Superset
  | Disjoint
  | Unknown
```
Representing a relation in this way is less precise because `subset`, `superset` and `disjoint` are not mutually exclusive. In fact when comparing two empty sets together their relation is all three. So instead I'll like to change the relation to be a bool for each of the properties.

An added bonus is it makes it simpler to implement `union`, `inter` and `complement`. I validated the current implementation quickly using Alloy and updated the model accordingly in this diff.

Reviewed By: viratyosin

Differential Revision: D63426623

fbshipit-source-id: 5e8be2c624eebf6d9b6f23d05f4fc58f403f56d5