Switch Typing_refinement to use ApproxSet.relate
Summary:
Currently `Typing_refinement` splits a type by transforming the type we are splitting (the scrutinee) to a datatype and the predicate to a datatype and computing the complement of the datatype. We then ask if they are disjoint to determine which side of the partition the type fall under.

Instead of computing the data type of the predicate and then computing the complement, we can utilize the `ApproxSet.relate` to get a `SetRelation.t` back. From there we can do the partitioning based on the relation.
* If the scrutinee is a subset of the predicate then it falls in the left partition
* If the scrutinee is disjoint from the predicate then it falls in the right partition
* Otherwise it spans both

Reviewed By: viratyosin

Differential Revision: D63430983

fbshipit-source-id: fac9a6f1be078f2f22bdf12a8a66aa47d56ec14d