Make `--ignore-hh-version` work even with .hhconfig changes
Summary:
check that previous diff has been released in watchman before committing this diff, or you'll break `--ignore-hh-version`. I also expect tests to fail until the previous diff is released.

Current situation: if .hhconfig has changes, there's not way to do a saved state init. This is because we query the saved state via watchman and watchman demands a non-empty project metadata. So currently, with `--ignore-hh-version` we'd ask prod hh to give us a project metadata so we can pass that to watchman... This hack unfortunately doesn't work if .hhconfig is modified because we'd get the wrong hash for the part of the project metadata corresponding to .hhconfig (even with prod hh).

The previous diff changes watchman so we can pass an empty project metadata to watchman, and that means "give me a saved state with any project metadata" (instead of "give me a saved state with project_metadata = ''").

So now, with `--ignore-hh-version`, we can instead pass no project metadata to watchman and stop querying prod hh. That way, `--ignore-hh-version` also works when .hhconfig has changed.

Reviewed By: vassilmladenov

Differential Revision: D62025125

fbshipit-source-id: a5b2eba4bb2c08f92eed075b07baedc854d8a15a