Type splat: subtyping
Summary:
By definition, a splat parameter is the last one in a function type. There are three rules for function subtyping. (Here we use `_` for meta-level ellipsis, and `...` for variadic and splat *syntax*.)

* If the splat parameter appears only in the subtype.
```
   (function(u1, _, um, ...u): s)
<: (function(t1, _, tm, tm+1, _, tn, optional tn+1, _, optional tp, t...): r)
```
holds if corresponding components preceding the splat are in a contravariant subtype relationship, and the remaining parameters in the right-hand-side are a subtype of the left-hand-side splat parameter when considered as a tuple type i.e.
```
ti <: ui for i<=m
(tm+1, _, tn, optional tn+1, _, optional tp, t...) <: u
s <: r
```

* Similarly, if the splat parameter appears only in the supertype.
```
   (function(t1, _, tm, tm+1, _, tn, optional tn+1,..., optional tp, t...): r)
<: (function(u1, _, um, ...u): s)
```
holds if
```
ui <: ti for i<=m
u <: (tm+1, _, tn, optional tn+1, _, optional tp, t...)
r <: s
```
* If the splat parameter appears in both, in the same position, just compare the underlying types, as with variadics:
```
   (function(t1, _, tm, ...t): r)
<: (function(u1, _, um, ...u): s)
```
holds if
```
ui <: ti for all i
u <: t
r <: s
```
In essence, a splat parameter is compared with a tuple type built from the parameters in the other type, with optional parameters mapping to optional elements of the tuple, and variadic parameter mapping to a variadic element in the tuple.

There is one case not covered above: if the splat parameter appears in both, but in *different* parameter positions. To deal with this we would need support for splat elements *in tuples*.

Reviewed By: enetsee

Differential Revision: D63316375

fbshipit-source-id: 4680d2ed804ab458f19b630f001fb8e44b0c371c