Internalize class to string conversion in ReflectionClass
Summary:
Very similar diff to {D54882879}

Internalizes the to-string conversion created by the `__init` method so we don't log on arbitrary WWW callers  while we restructure the NativeImpl and associated structures. For example, when {D58398299} lands it will be better to preserve its output in the ReflectionClass constructor rather than convert it.

Reviewed By: DavidSnider

Differential Revision: D63649810

fbshipit-source-id: 0c7c48f98f78094440b49228dbbac1e50adbe6a5