Fix parse_ast and remove parse_and_mutate_program
Summary:
`parse_ast` does only partial semantic analysis (just type resolution) which contradicts its intent and comment:

```
/**
 * Parses a program, performs semantic analysis and returns the resulting AST.
 * Diagnostics such as warnings and errors are reported via `diags`.
 */
```

It also doesn't correctly pass parameters to the created `sema_context` object. Fix this and remove `parse_and_mutate_program` which is now equivalent to `parse_ast`, just with a messier API and a second overload that does something subtly different.

This reduces the public API surface and makes users fall into the "pit of success" by getting a valid AST by default. It is particularly important for initializers which are pretty much unusable without type inference performed by the semantic analyzer.

It uncovered bugs in tests (invalid Thrift code) - fix them.

Add an option to skip lowering of structured annotations. Lowering is only needed for legacy plugins that don't understand structured annotations, so setting it will be moved to plugins in a follow-up diff. It is not done in this diff to keep it manageable and to avoid conflicts with other compiler changes in flight.

Reviewed By: thedavekwon

Differential Revision: D63659880

fbshipit-source-id: d2acc168e1decd5a8dd6fcce421252ff90f27315