Ensure that InputStreamByteBuf terminates if end of stream is reached
Summary:
In MQTT, we're experiencing extremely high CPU in what appears to be an infinite loop in `MqttThriftHeader.read0` method.

P1623614975
```
 --- 57497585148 ns (37.58%), 574638 samples
  [ 0] java.io.ByteArrayInputStream.read
  [ 1] com.facebook.thrift.util.InputStreamByteBuf.read
  [ 2] com.facebook.thrift.util.InputStreamByteBuf.readInt
  [ 3] com.facebook.thrift.util.InputStreamByteBuf.readFloat
  [ 4] com.facebook.thrift.protocol.ByteBufTCompactProtocol.readFloat
  [ 5] org.apache.thrift.protocol.TProtocolUtil.skip
  [ 6] org.apache.thrift.protocol.TProtocolUtil.skip
  [ 7] com.facebook.swift.titan.iris.protocol.MqttThriftHeader.read0
  [ 8] com.facebook.swift.titan.iris.protocol.MqttThriftHeader$$Lambda$1782/391454626.read
  [ 9] com.facebook.thrift.util.SerializerUtil.fromInputStream
  [10] com.facebook.titan.iris.swift_serialization.SwiftSerializer.unserializeInputStream
  [11] com.facebook.titan.iris.swift_serialization.SwiftSerializer.unserializeNoHeader
  [12] com.facebook.rtgw.handlers.delivery.RealtimeDeliverySender.sendRealtimeDeliveryRequest
```

Note that `MqttThriftHeader` does not have float types, it has two strings, so readFloat indicates bad input data in this case.  All threads appear to be spinning in this method. The only espace from `read0` is if we get a and end struct byte.

```
while(true) {
            TField __field = oprot.readFieldBegin();
            if (__field.type == 0) {
                oprot.readStructEnd();
                return builder.build();
            }
```

End struct is read in `readFieldBegin` with the following code
```
    byte type = readByte();

    // if it's a stop, then we can return immediately, as the struct is over.
    if (type == TType.STOP) {
      return TSTOP;
    }
```

For this implemenation readByte() is implemented as follows

```
  public byte readByte() {
    read(Byte.BYTES);
    return buffer[0];
  }

...

  private void read(int length) {
    try {
      inputStream.read(buffer, 0, length);
    } catch (IOException e) {
      throw Exceptions.propagate(e);
    }
  }
```

When the end of an InputStream is reached, the return value will be -1. If this is the case, `read(int length)` returns successfully and `readByte()` returns `buffer[0];`. Buffer will be left over data from the previous read. In our case it appears the `buffer[0]` value get's coerced in the `Type.FLOAT`. This triggers the `read0` method to enter the `skip` method and continue reading. All future reads result in the same data from the last successful read, causing  an infinite loop. We believe this is malformed data from some bad client that MQTT is talking too.

This diff adds a boundary condition to ensure that if we attempt to read past the end of the `InputStream` that we throw an exception. This is roughly equivalent to how the `AbstractByteBuf` implementation works, where it checks to ensure that when `readByte()` is called that the bytes actually exist.

```
    public byte readByte() {
        checkReadableBytes0(1);
        int i = readerIndex;
        byte b = _getByte(i);
        readerIndex = i + 1;
        return b;
    }
```

Reviewed By: robertroeser

Differential Revision: D63777839

fbshipit-source-id: 0971326a51766fa4a508843398a7b895cafda3f4