Allow root to implicit imply the naming table
Summary:
`hh_single_type_check` has a nice functionality that given a `--root`, it can find a naming table (though somewhat arbitrarily chosen if there are multiple).

This was artificially not allowed for `--get-some-file-deps` by a premature option validity check. This diff changes it, so that it can benefit from it.

Reviewed By: viratyosin

Differential Revision: D63715607

fbshipit-source-id: a49dd35ec20e9d0ec2e75d6743d0fc0460099675