strip extra leading / from multifile file names
Summary:
In a multifile test, every file must begin with four `/`, eg:
```
//// foo.php
```
The number of leading `/` was not enforced and the parsing code added the extra `/` to the filename.  So, given:
```
////// foo.php
```
the filename was the string `//foo.php`.  This leads to all kind of hard to detect problems when writing multifile tests for PackageV2.

This diff strips the extra leading `/` characters from the multifile file names.

Reviewed By: DavidSnider

Differential Revision: D63634767

fbshipit-source-id: 97b0af4e304e5433c6008a600b77ad1b3fd9466d