Add BDD Implementation of ApproxSet
Summary:
# Context
The current implementation of `ApproxSet` is represented using [Negation Normal Form (NNF)](https://en.wikipedia.org/wiki/Negation_normal_form). This representation is simple to build, but is more challenging to reason about. For instance the current implementation isn't able to determine reliably that two sets are equal to each other, or that the union of a set with its complement is equal to the universal set.

I've experimented with other normalizations such as DNF or CNF, but the size of these normalizations can grow exponentially with the number of operations performed on them.

The goal of this diff is to provide a better implementation of `ApproxSet` that has acceptable performance without making the algorithm too complex.

# Solution

While looking for solutions for this class of problems I learned of [Binary Decision Diagrams (BDD)](https://en.wikipedia.org/wiki/Binary_decision_diagram). Their structure allows for representing semantically the same sets in a consistent way. I implemented this data structure and used it to produce an alternative implementation of `ApproxSet` that matches the API called `BddSet`.

The implementation itself is comparable in size to `ApproxSet` requiring roughly 100 extra lines of code (including the BDD implementation). With that we significantly increase the completeness of the `relate` function as seen in the new `test_quick_complement` test case added to `Approxset_test`

Performance was comparable with the existing implementation. NNFs grow linearly with the number of union/intersection algorithms, while BDDs grow with the number of atoms. The number of nodes could grow more than linearly with the number of atoms, however there are known heuristics that lead to better performance, particularly how you order elements. The simplest is to choose an ordering of atoms that aligns with subsetting. Stated plainly we want an order such that `A ⊇ B implies A < B`, where `<` is the order relation for the BDD. This means we want to keep larger atoms closer to the root of the BDD, which corresponds to lower in the ordering.

With this choice, the test suite was within 75% performance of the previous implementation. This is with a simple BDD implementation that doesn't try to do any optimizations. Since the size of the BDDs we are building will typically be small (fewer than a dozen nodes on average, with extreme cases being less than hundred), I figure this implementation will be good enough. If perf does become an issue there is more optimizations and caching that can be added. I'll discuss this in more detail in a subsequent diff when I update `Typing_case_type` to use the new implementation (at which point I will do a perf test).

Differential Revision: D63506263

fbshipit-source-id: 4be832edde56f107bf93c9c3bb5e1e3513412585