Do not ignore Nameof token
Summary:
The nameof token being an IgnoredToken was causing the expr node to have the wrong position; it would just merge into the target pos.

The decl ext uses this position to slice the source  text as its generic fallback, so you'd get back something that looks like a global const.

Reviewed By: nt591

Differential Revision: D63842284

fbshipit-source-id: 294ca5b8275fa0f23df75ed9f14ebbfd2eba111a