Convert the remaining few tests to work with CheckBuiltinParamTypeHints=2
Summary:
Some of these tests are not really needed anymore, but were written when
security vulnerabilities were fixed due to incorrect object types received by
extensions, so keeping them.

I guess the proper type checking of builtin params avoid this class of
vulnerabilities altogether.

The only 2 tests still modifying CheckBuiltinParamTypeHints are the ones that
test for the correctness of the feature, they will be removed together with the
feature.

Differential Revision: D63872506

fbshipit-source-id: 0f05e2cda2795850eeaf9ac5883bb2e591e23ecd