Make newtype treatment same as alias
Summary:
We were generating an indirect inabitant for `newtype`s through a function call. This was to think a bit ahead when we'll partition definitions into multiple files.

I decided that it is a bit premature and complicates some of the implementation with respect bounds. So I'm changing the treatment of `newtype`s to that of `alias`es.

Reviewed By: andrewjkennedy

Differential Revision: D63690322

fbshipit-source-id: 6b12b0634e6a21b2bbb9ecf095e89c742c6a74fc