More special casing for solved tyvars
Summary:
In legacy errors the reasons for tyvar solutions _replace_ the reason for the tyvar. In extended reasons we record the solution to ensure we don't destroy provenance. To recreate the legacy error messages we want to:
1) Propagate solutions through `Flow`s and use the solution reason when we find the corresponding tyvar
2) If a `Flow` doesn't contain a tyvar, use the outermost reason.

This diff special cases the situation where we have a `Solved` reason wrapping a flow which doesn't contain a type variable

Reviewed By: madgen

Differential Revision: D63898873

fbshipit-source-id: fdeb515fdaa64fec728783d6a0ff7e2a92edf964