Adds test to exercise bug when rendering supertypes with Flows
Summary:
Extended reasons accumulate provenance on the subtype during subtyping but we do accumulate provenance on the supertype during typing, resulting in a Flow. At the moment we aren't handling this meaning that we end up pointing at the outermost rather than innermost reason, which is what legacy reasons will do.

This case wasn't exercised by existing tests so we add one here

Reviewed By: madgen

Differential Revision: D63900544

fbshipit-source-id: eca81a69bffbc522c6ecbb7e81c5ede2cd760d5c