Propagate solutions rather than rendering
Summary:
At the moment we always print the solution to a type variable when rendering the reason of a type appearing in the claim.

This is inconsistent with legacy errors which always capture the outermost reason - we replicate this by recursing in the outermost reason in a `Flow` but `Solved` currently ignore this.

This diff changes the behavior so that we propagate solutions but only use them when the outermost reason corresponds to a type variable

Reviewed By: patriciamckenzie

Differential Revision: D63829943

fbshipit-source-id: ea28687016e97eb32d5449a39f7fdd3e1314d8d4