Reverse top-level Flows in supertype
Summary:
Although we accumulate provenance on the subtype during subtyping, we actually record typing time provenance on both types so the supertype may be a `Flow`.  In this case we actually want the inner most element of the flow rather than the outer most (which is what we want for the subtype)

This diff adds a function to reverse top-level flows and applies it to the supertype when rendering a subtyping error.

Reviewed By: madgen

Differential Revision: D63900892

fbshipit-source-id: 2c2034476c2ae0b5e68c01622f18bb9de8ecf090