Consistently record provenance on lower bound
Summary:
Extended reasons usually accumulates provenance on the subtype except for one case: when adding an upper bound to a type variable. This oversight was causing reasons to be dropped when we have an return type containing an invariant type variable.

This diff modifies provenance generation so that we always accumulate on the lower bound during the transitivity step and makes the necessary change to the derivation generator to recognise nested lower bounds as a transitivity step.

Reviewed By: patriciamckenzie

Differential Revision: D63749923

fbshipit-source-id: af0f87c65e345721e824a33c52d0b8632a6762b1