Fix like nullable type treatment in the emitter
Summary:
When like type support was implemented, we didn't correctly mark like types as not being "simple" types. This led to nullable  part of the type being swallowed.

This diff fixes that by adding an arm and also gets rid of the catch all arm of the match expression so that this doesn't happen again in the future.

Reviewed By: mdko

Differential Revision: D63980235

fbshipit-source-id: 54b5ab697452a1e761340380a054b93cc93a4e09