Fix a bug related to literalsToRemove and add a couple more asserts
Summary:
Since we pooled the literals and separated them their corresponding code
sequences (D7622217), the literals and their code sequences are not subject to
the same alignment constraints.  However, the code was still avoiding the
process a literal inserted in literalsToRemove if it was subject to alignment
constraints.  That was incorrect because, a literal is inserted into
literalsToRemove when we know it won't be needed, so we should remove them
regardless of alignment.  This diff also updates a misleading comment and adds a
couple of asserts.

Differential Revision: D63992491

fbshipit-source-id: 698c795e2dca0daa7464aaecd1e19cf83b742119