let RegexMatchCache compress regex-sets as dynamic bitsets
Summary:
It is desirable to avoid the excess memory usage in the match-to-regex-cache and in the string-queue-forward from directly storing sets of regex-pointers.

It is typically expected, in the intended use-cases, that there be few-enough regexes such that iterating the set bits of a bitset would not be too expensive in practice, even though it is linear in the number of bits and not in the number of set bits.

This is a simpler implementation of a dynamic bitset.

Reviewed By: mdas7

Differential Revision: D63338697

fbshipit-source-id: 48dc07f216306ecabcff72feef3ef6a3d8b8602f