Typing_defs::typedef_type - combine visibility and assignment into single field
Summary:
Case types are implemented as typedefs with a special visibility. They fit as a "normal" type alias by creating a union of their variants to use as the assigned type.

Case types with where clauses will need to maintain more structure and information.

The approach pursued by this diff is to make the assigned type either a "simple type" or a "case type". I merge the visibility into the "simple type" variant since the visibility can imply the assignment variant.

I tried turning the visibility enum itself into an ADT but ran into issues with oxidize.

Reviewed By: dlreeves

Differential Revision: D62251888

fbshipit-source-id: 0642d86520cba001f85fccd44df2774e66d1693d