Simplify generator directory structure
Summary: The current generator directory structure is quite confusing with the C++ generator logic spread across `gen` and `generate`. Other generators don't use the `gen` directory at all and even in the C++ case it is only used for two things, name resolver and reference type. To simplify this, merge the content of `gen` into `generate` and remove the former.

Reviewed By: yoney

Differential Revision: D63987934

fbshipit-source-id: 79e9c87835f9be37ed3a4305d92d42dd827ffe83