Implement HTTP/2,3 padding frames
Summary:
Use reserved, unused HTTP/2 and HTTP/3 frame types as padding. We want these frames to
be ignored/discarded by the other endpoint of the connection and not change any aspect of
processing.

== Concerning the safety of this approach

RFC 9113 (HTTP/2) on using reserved frames for padding (section 4.1):
> Implementations MUST ignore and discard frames of unknown types.

RFC 9114 (HTTP/3) on using reserved frames for padding (section 10.7):
> HTTP/3 can [...] employ the reserved frame and stream types [for padding].

Note there is an errata in RFC 9114 section 10.7: it currently states:
> Where HTTP/2 employs PADDING frames [...]

HTTP/2 does not define a PADDING frame. However, in this diff, we are suggesting to use
the currently unused and reserved frame type `0xbb` as a PADDING frame in HTTP/2.

Reviewed By: hanidamlaj, knekritz

Differential Revision: D62032257

fbshipit-source-id: 8abf1009c11fc9bca2829a3233c15289381715b1