Move like type hints into experimental feature mechanism
Summary: Add a new `like_type_hints` experimental feature, and only allow `~` in type hints when the feature is enabled. Remove all of the custom checking from (typechecker) elaboration,, and the `like_type_hints` configuration option. This means that hackc/hhvm now need to enable the feature to allow `~` in a hint, whereas before it was always allowed as the check was only in the typechecker. This necessitated updating a lot of test files.

Reviewed By: madgen

Differential Revision: D63987282

fbshipit-source-id: 0da8a70e16e08797c0856c99afc6d13d9d2f523c