Do not run jit for invalid functions
Summary:
After a JIT request is enqueued, the corresponding function may be destroyed before it is picked up by a JIT thread. Check if the function id is valid before jitting because it is set to invalid before treadmilling a function.
More subtly, the unit or the class corresponding to the function may be treadmilled and the function may be picked up for jitting before the function id itself is set to invalid. In order to prevent such scenario, we mark all functions of a unit/class as zombies before the unit/class is treadmilled and refuse to jit such functions.

Reviewed By: paulbiss

Differential Revision: D63151246

fbshipit-source-id: 9aca85b7a489742832ea194bb041bf29385a42cf