Record case type bound subtyping
Summary:
Case type generation wasn't recording the fact that adding bounds on case types establish a subtyping relationship.

This diff adds support for that.

This make non-termination of a number of predicates including the disjointness weakening and picking a subtype. For example,
```
case type CT as C = C;
class C {}
```

Recursive predicates chase recorded subtypes, so there is infinite recursion that follows `CT -> C -> CT -> C -> CT -> ...`.

To overcome the termination problems, we switch to monotonically growing sets in type weakening and a step-wise stochastic walk in the subtyping relation.

Reviewed By: CatherineGasnier

Differential Revision: D64113991

fbshipit-source-id: 1f322eeaef7ddf8e6e4262226ff61d55df6b689c