Simplify type weakening and subtyping
Summary:
We had two modes of the `subtype_of` relation. One that non-deterministically picks an inhabited type and another that computes a list of subtypes.

The latter is expensive and it leads to termination problems down the line.

Luckily, the latter was only needed for the type weakening predicate that is used for case type disjointness. I noticed that this can be more efficiently implemented without using the `subtypes_of` predicate.

This diff eliminates the mode of `subtypes_of` where it returns all known subtypes and keeps the usage that picks a single inhabited one.

Reviewed By: CatherineGasnier

Differential Revision: D64080718

fbshipit-source-id: 0764050145b906bc7b09d52b3bf18abd8674ef89