Produce eventually inhabited subtypes
Summary:
The `subtype_of` relation was restricting the produced subtypes to immediately inhabited ones. This is necessary when it is used for expression generation, but not so much for type hint generation.

This diff relaxes the requirement so that the type generator generates eventually inhabited types instead of immediately inhabited ones.

This gives us more varied generated programs.

Reviewed By: CatherineGasnier

Differential Revision: D64114005

fbshipit-source-id: 96465ea420a5d7ab1ec2c0c4bb2e2a83012bd44f