Compile C::class shape keys to SFclassname
Summary:
D59345761 introduced SFclassname, a place to hold `nameof Foo` shape keys and some of the typechecker `SFclassname -> TSFclass_const` boilerplate.

This diff does the rest of the TC implementation. Unlike in the runtime where we focus on being a string, the goal here is to model `Foo::class` keys as much as possible and their type `TSFclass_const`, so that definition `shape(C::class => int)` and indexing `$s[C::class]` may be updated gradually and independently.

I believe that by migrating, and banning XHP class name keys -- due to the underscore conversion -- and modeling the runtime's lead namespace stripping `(\A\B -> A\B)` we can transition to `SFclassname` becoming a `TSFlit_str`, allowing the keys to coexist with string keys and each other.

Miscellaneous:
- inlined the shape get_name() and get_pos() impls since they're just used for that one sketchy check
- Tagged the elaboration of self::class keys as dead code (see `typecheck/nameof/shapes_illegal.php`)
- killed a few more wildcard matches

Reviewed By: viratyosin

Differential Revision: D64015310

fbshipit-source-id: cba0e8c18c8fa0f91aae8a6837f60578b3c6e05a