allow invoking private static methods of the required class
Summary:
If a trait as a `require class C;` constraint, then it can invoke all `private` methods of `C`, including the static ones.

LSB is not an issue because:
- if a trait `T` has a `require class C` requirement then class `C` is final (enforced by Hack and HHVM)
- trait `T` will only be used by class `C` (enforced by Hack and HHVM)

These imply that at run-time, inside trait `T`, `self` always resolves to `C`.  Additionally it is now an Hack error to invoke a static method on a trait that has a `require class` attribute, so HHVM will always have the class context when it resolves `self`.

Reviewed By: andrewjkennedy

Differential Revision: D64173722

fbshipit-source-id: daa190a2df33143bfacfae7919c297427ec758c8