let Snapshot::forEachSetting take FunctionRef
Summary:
The function `folly::settings::detail::SnapshotBase::forEachSetting` is virtual in an abstract base class. But it takes a callback function, which means it cannot be a member function template. The options are restricted to a type-erased function-wrapper or another abstract base class with a virtual member function for the callback. It currently takes `std::function` as its type-erased function-wrapper.

But `folly::FunctionRef` is likely a better-performing fit: it may be passed in registers on some platforms, it does not require copying callback-object state, and it does not require heap-allocation ever.

Differential Revision: D64237891

fbshipit-source-id: 22e7a51e7c91c8293c1ca839b0bcced7a82be492