Check where clauses in case types
Summary:
This diff updates subtyping to consider where clauses for case types (experimental feature).

When checking if a type T is a subtype of a case type with where clauses, T must satisfy at least one of the case type variants where "satisfy" here means that T is a subtype of the variant's hint and that all of the variant's where clauses hold. So, for each variant, we conjunct the proposition with the hint with the propositions of the where clauses.

Backtracking should not be an issue since the variants' hints must still be disjoint.

Reviewed By: vassilmladenov

Differential Revision: D64077762

fbshipit-source-id: 2c7dd29dca822ae4798eb81f43c8ecf344519dfd