Recursive types: improve error message when error occurs due to recursive type definition
Summary: Currently whenever we localize a recursive type alias, whenever we encounter the recursive use of the type, we produce mixed. This results in confusing error messages mentioning `mixed` without explanation of where that `mixed` comes from. This diff fixes that by introducing a new `Typing_reason.witness_decl`. See test case expected files for the before and after of the errors.

Reviewed By: enetsee

Differential Revision: D64331485

fbshipit-source-id: dc87874928bf00f4b1f9ff0d62754537809911e7