Fix bad regex detection
Summary:
S458691 didn't have hhvm_configtool run, but even if it did the validator fails due to minor differences in hhvm regex and rust hhvm.

Rust doesn't require regex pattern boundaries so one doesn't need to escape the boundary character

We force configs to use the boundary character '/' and strip it out so if we see '/' preceded by a valid tw job char a-zA-Z0-9_- we can flag it as unescaped.

Differential Revision: D64367016

fbshipit-source-id: a9bd76dddfe2319591d702a5203a63f797b75ada