More comprehensive test suite
Summary:
As the variety of programs we generate grow, it becomes harder to cover all cases.

This diff changes the well-typedness tests so that they can take a start and end seed range for which they run the tests in parallel.

## Why not increase the sample size?

Tests run in stress test mode to be eligible to run on CI. So each separate test needs to be kept under a certain duration. Separating into separate tests each with 100 samples help with that.

Reviewed By: CatherineGasnier

Differential Revision: D64267706

fbshipit-source-id: ee0f8a67a8906182be7afc0905b1c0383fea8637